[connman] Fix dev mode plugin firstboot issue. Contributes to JB#45772
This fixes the issue of not having developer mode in some cases when
doing first boot with USB cable plugged. For some reason the interface
is not IFF_LOWER_UP but the change is passed as change.

Added new function to check if the device exists (TODO make a new name
for "pending_device*") and it is used when adding, getting new and del
notifications from kernel netlink. Made improvements in management of
the developer mode.

Also reduced developer mode plugin debugging noise. Change logging from
negative logging to positive and more informative logging in order to
reduce noise in logs when debugging.