technology: Implement functions for pre and post setup in user change
When doing preparation setup all technologies in use will be disabled.
Offline mode is not changed and does not affect the preparation setup.

When doing post setup the offline mode is read from the new settings
file. Next each technology is updated based on its "Enable" value in the
settings file. If the technology is to be enabled but ConnMan is in
offline mode enabling of the technology is skipped.