login manager: Implement login manager stub and config option for it
[connman] Implement login manager stub and config option for it. JB#49403

Add a stub for login manager implementations. This is to be executed
after storage but before everything else in order to make it possible to
react to active user also if connmand crashes. This is important since
initializing storage does not load anything and user can be changed
without service unloading in case of a crash. Using login manager as
plugin would result in first using the effective user running connmand,
then when plugin is loaded and user is checked an unnecessary change
will happen. This stub is place for adding initialization for specific
login managers which depend on the build time configuration, e.g.,
systemd.

Add EnableLoginManager boolean config option to main.conf. Setting
value to true allows to use login manager, false ignores login manager.