login manager: Implement login manager stub and config option
[connman] Implement login manager stub and config option. JB#49403

Add a stub for login manager implementations. This is to be executed
after storage is initialized but before everything else in order to
make it possible to react to the current active user also if connmand
crashes. This is important since initializing storage does not load
anything and the user can be changed without service unloading in case
of a crash. Using login manager as plugin would result in first using
the effective user running connmand, then when plugin is loaded and the
current active user is checked an unnecessary unload-load process will
happen. This stub is place for adding initialization for specific type
of login managers which depend on the build time configuration, e.g.,
systemd.

Add EnableLoginManager boolean config option to main.conf. Setting
value to true allows to use login manager, false ignores login manager.
By default, login EnableLoginManager is off and needs to be explicitly
enabled.