systemd login: Implement systemd login manager
[systemd login] Implement systemd login manager. JB#49403

Implement systemd login manager using libsystemd with GIOChannel fd
watcher. The systemd login manager has a state machine documented in the
source file. States are tracked one step backwards (i.e., old state) to
ensure that transition is correct. This will aid in future development.

The systemd login manager registers to systemd to listen for login
changes based on "session". When there is an event in the systemd login
fd a status check is initialized as a delayed check (100ms delay). If
there is already a pending check new one is not started. Every status
check removes itself after doing the active user check on systemd login.
The check uses "seat0" as the default seat to get the session status and
uid.

In case the user's state is "active" the user change is done using
__connman_storage_user_change() using the active user or when the user
is loggin out (state is "logout") user is changed to root user (uid =
0). This is done to make it possible to hide (unload) all previous user
services in order to ensure that the new user wouldn't be able to access
them (and to see, e.g., WiFi service passwords).

The user change request is always pending for a reply from vpnd and
while waiting for the reply a new status check is not made. In case
there is a timeout error sent back new status check is initialized.
In case of a timeout the active uid is not changed and the new status
check initiates new user change process if the user is still in
"active" state.

The delayed check will get the most recent information when it is
changed. Having it delayed has both up and downsides: upside is that not
every fd action is a session change that requires connman action, downside
is that some session changes may come at fast pace and in theory these may
be missed. But the latest state is still the utmost intrest and this is
achieved with the delayed check.

The connection to systemd login is attempted to be restored if the fd is
detected to be closed by the GIOChannel watch. In this case both
GIOChannel and the systemd login connection are re-established, with a
500ms delay between the attempts until success. However, failure to
connect to systemd login may yield bigger problems and is not dealt in
much finesse here by anything else than a mere re-connection attempt.