storage: Add internal function for changing user
[storage] Add internal function for changing user. JB#49403

Implement __connman_storage_change_user() that is similar to the D-Bus
method function with small differences. Appropriate error is returned
and -EINPROGRESS means that the call to vpnd was made and answer is
pending. The pending message is omitted and this is handled now in
change_user_reply().

This function is to be used by login manager listerners. With callback
the result of the operation will be returned back to the caller, when
D-Bus calls finish.