storage: Implement preparation of user change and improved change cb
Implement a preparation mechanism for user change to be done in connmand
only. This is initiated at early initialization stages of connman which
only changes the paths to the user paths of the currently logged in
active user. No preparations, unloading, loading or post actions are
done because the components that implement these (technology.c,
service.c etc.) are not initialized yet.

The idea is to do the process in reverse when preparing, connman sets
first the paths then communicates with vpnd. The preparation is not
necessary for the vpnd as it can process the D-Bus method call only
after it has been initialized - by systemd. Thus, everything is set when
the user change message is processed.

When preparing the user change callback is called twice, first with
-EINPROGRESS to indicate that change has been made and pending for vpnd
result. Then when vpnd returns it will be called again. If in any of
these steps there is an error it will trigger setting the user back to
user "root", so default storage location is used. In this case the user
change cb is called with uid 0.