[unit] Implement unit tests for systemd login. JB#49403
Implement unit tests for systemd_login.c. Use mainloop and context to
simulate glib mainloop in the tests. This way the delayed callback in
user change can be simulated as if the reply would be coming from vpnd.

Created mocks for glib main loop iteration using a specific poll
function to fool GIOChannel to think there is data incoming. This allows
to simulate the systemd login fd indication that there is some change in
user session.

The idea is to create the main loop and use the context to iterate
through events, some of which are triggered by the polling function for
the systemd login fd, and some are timeout functions added by the test
and the systemd_login.c. Both of these are regarded as events and in
some cases, when the main context iteration is set to not to wait for
events they are required to be dispatched manually. It may not be 100%
accurate as all the events dispatched may not be possible to get with
the public API.

Use always expected event count + 1 to test that no additional events
are done. Three different iteration amounts causing 0.120s, 0.300s and
0.600s iteration durations are used. If either, the maximum number of
events given or the iteration count is reached the main loop simulation
is stopped. The main loop simulation does a 20usec sleep between the
loop iterations.

Implemented log hook to check the return values output to log to check
that the change is a success.

Implemented also 4 full tests to test first the user change as normal
and then getting simulated notifications about user changes.