storage: Use uid and user data in finalize.
[storage] Use uid and user data in finalize. JB#49504

Implement connman_storage_update_finalize_cb() to allow setting of the
finalize callback to user change after initialization. This function is
called last in the process.

It is useful for the vpn plugin in connmand to set a callback for
requesting the VPN connections from the vpnd. The vpnd could send the
signals after the VPNs are loaded from keyfiles but it will come too
early to connmand. It may be that the D-Bus signals are processed before
the user change is completed. Allowing to register a callback to be
executed and to explicitly request the VPNs from vpnd is a better, more
synchronized solution for connmand.

In addition, removed the use of username in the finalize callback and in
storage.c alltogether. Use of uid_t is sufficient.