vpn: Use get_connections() as connmand finalize func.
[vpn] Use get_connections() as connmand finalize func. JB#49504

Implemented vpn_finalize() function to register as callback to storage.c
to be ran as the last function in the user change process. This function
simply calls the get_connections() as if connmand notices vpnd being up.

The purpose of this is to get the users' VPN services from the vpnd when
connmand has dealt its part in the user change process. Since the
message to vpnd is sent first, vpnd should have loaded all VPNs when
the request sent by get_connections() arrives.