systemd login: Disable all technologies when user is opening session.
[systemd login] Disable all techs when user is opening session. JB#49436

When user login is changed via systemd login monitor in some occasions
root user, the default system user, may have technologies enabled which
may be confusing to the user and system components when the user is
opening up session. This change ensures that prior to user change
happening after the session is active the technologies are disabled.

This allows some technolgies, such as rfkill enabled WiFi, to be put
down in time before it is enabled again by the user configuration. This
can be the case in both boot and user change. The reason is that when
user is logging out root user is enabled, and the technologies enabled
on the root user are thus enabled as well.