openvpn: Improve config data processing
[openvpn] Improve config data processing. Contributes to JB#55889

Improve how the config data options are processed in OpenVPN. Add a
config option type that is providing better granularity in how the
options should be processed when adding to the task, which eventually
sets them to the binary.

This now allows to add string variables that have a default value when
they are used alone by setting the "opt_type" to OPT_STRING and setting
the "has_value" to 0.