openvpn: Handle script context in notify
[openvpn] Handle script context in notify. JB#55889

Add handling of the script context in OpenVPN notify function. The
script for OpenVPN sends also the context provided by the binary and it
is to be handled prior to the environment values as a distinct
indication whether initialization or restart is taking place. In case of
any other than "init" propagate ECONNRESET error through the callback to
vpn-provider.c and disconnect the VPN. This is due to the fact that vpnd
should be in charge of reconnections and by doing this the errors can be
recorded for future use.

Having uncontrolled restarts had issues with TCP and a misconfigured
connection using UDP may also reset itself with SIGUSR1 and repeditely
call the up script which would lead up piling of new routes. UDP does
not have such option to prevent connection retries so in such case
processing context works. User is also capable of preventing this by
remapping the SIGUSR1 to something other but this approach is a fallback
in case the SIGUSR1 is still used with OpenVPN.

This was seen to happen with some UDP VPNs that use and require
--comp-lzo, which was omitted from options as an error (and fixed in
this patch-set), and then the binary reports:
  write to TUN/TAP : Invalid argument (code=22)
and as a result, after the ping-restart timeout has been reached,
OpenVPN does an internal restart and informs it via up script.