[nat] Add IPv6 preparation and double NAT support. JB#52923
Implement prepare and restore functionality to for IPv6 NAT. It was
required to add IPv6 type nat struct that is not enabled when nat is
reqularly enabled to maintain the data. The support is enabled by:
 - support setting IPv6 forward for all interfaces
 - support setting accept_ra for the transport interface via ipconfig
 - support adding firewall forwarding rules for the both interfaces
 - support setting ndproxy as enabled with neighbour proxy via inet

Add also support for double NAT. This is done by adding a second
set of NAT addresses with an additional destination address using the
interface as the MASQUERADE rule out interface. When tethering enables
NAT this new rule is enabled as well since they both are for IPv4. If
the double NAT is enabled when NAT is already running similar restart of
just re-adding the iptables nat table rules is done as it previously
was.