[clat] Re-add prefix query timeout, fix prefix change check. JB#52923
Always re-add the prefix query timeout function after the query has been
done. This should make it simpler for adapting the TTL-10s requirement.

Also limit prefix length to 120, this might need to be changed later but
with prefix lenght of 128 the address space would not be able to handle
additional clat interface address.