[nat] Keep ref to IPv6 ipconfig for restore, fix error mem leaks. JB#52923
The restore is important to be done when IPv6 setup is not needed. In
some cases where the component/plugin using this functionality uses an
external process which takes time to shut down cleanly and during this
time a service may have removed its ipconfig making restore impossible.
Thus, nat must keep the reference to the given IPv6 ipconfig for the
duration the IPv6 nat is required. This makes it safe for the plugin,
for instance, to keep the ptr to the ipconfig as well between nat IPv6
prepare and restore.

Also make the free'ing of struct connman_nat a bit better. Using the
common free'ing function eliminates the existing memory leaks in case of
errors.