[clat] Support VPN over CLAT, fix tracked service setting. Fixes JB#60276
Add a specific VPN mode to CLAT. In this mode the default route on CLAT
device is dropped and in case tethering is or was enabled double nat
will not be set. Thus, tethering should work solely over VPN in such
scenario.

Change to use ipconfig.c IP address check. Plugins should not need to do
this extra.

Use the proper nat.h path for including it. The header is now included
in the Makefile.am properly.