[provider] Do not disable IPv6 if transport has only IPv6. JB#60276
In case the transport of a VPN does have only IPv6 it should not be
allowed to be disabled. This is a special case that can happen with the
presence of DNS64 where all traffic is tunneled over IPv6 using a
specific plugin and, thus, transport may not have IPv4 enabled. If
IPv6 would be disabled in this case the VPN naturally would not work.