[ipconfig] Revert upstream commit 79e3ee298. JB#60276
The commit added in a single slot manual IPv4 address to use system
entry. It does not seem to be used as such anywhere else like this but
instad used along with fixed method. This blocked sending of the proper
IPv4 data in D-Bus PropertyChanged signal in cases the method was set to
manual.

commit 79e3ee29897214150ca8fb4e4a9caaaa867c7c94
Author: Patrik Flykt <patrik.flykt@linux.intel.com>
Date:   Fri May 24 14:38:01 2013 +0300

    ipconfig: Use system set addresses for IPv4 property in 'manual' mode

    Addresses and gateway for the service IPv4 property being used in the
    system are available from ipconfig->system when the property is set to
    'manual'.