gdbus: Fix client mem leaks in case of errors.
[gdbus] Fix client mem leaks in case of errors. JB#51243

In any other case the msg was unref'd. Unref it also if the
method_call_data allocation fails.

Also call the destroy function for the data passed to the
g_dbus_proxy_method_call() if D-Bus message sending fails. Each use in
the bluetooth plugin does rely on the data being somehow deleted and the
return value does not seem to indicate freeing the data otherwise.